RF: Integrate initial tests for Stelzer algo.
